Rename single function files to the name of that function
