+++ src/server/lib/APIFetch.ts
@@ -1,7 +1,7 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
 
 import fetch, { RequestInit, Response } from 'node-fetch';
-import { getConfiguration } from '@/server/lib/configuration';
+import { getConfiguration } from '@/server/lib/getConfiguration';
 
 const {
   idapiBaseUrl,
