+++ src/server/lib/getAssets.ts
