+++ src/server/lib/middleware/index.ts
@@ -3,9 +3,9 @@ import cookieParser from 'cookie-parser';
 import compression from 'compression';
 import { helmetMiddleware } from '@/server/lib/middleware/helmet';
 import { loggerMiddleware } from '@/server/lib/middleware/logger';
-import { applyRoutes } from '@/server/lib/routes';
+import { applyRoutes } from '@/server/lib/applyRoutes';
 import { csrfMiddleware } from '@/server/lib/middleware/csrf';
-import { getConfiguration } from '@/server/lib/configuration';
+import { getConfiguration } from '@/server/lib/getConfiguration';
 import { serverStateLocalsMiddleware } from '@/server/lib/middleware/serverStateLocals';
 import { mvtIdMiddleware } from '@/server/lib/middleware/mvtId';
 import { abTestMiddleware } from '@/server/lib/middleware/abTests';
