+++ src/server/lib/renderer.tsx
@@ -5,9 +5,9 @@ import { StaticRouter } from 'react-router-dom';
 import { Main } from '@/client/main';
 import { brandBackground } from '@guardian/src-foundations/palette';
 import qs from 'query-string';
-import { getConfiguration } from '@/server/lib/configuration';
+import { getConfiguration } from '@/server/lib/getConfiguration';
 import { RoutingConfig } from '@/client/routes';
-import { getAssets } from '@/server/lib/assets';
+import { getAssets } from '@/server/lib/getAssets';
 import { getDefaultServerState, ServerState } from '@/server/models/Express';
 import { FieldError } from '@/server/routes/changePassword';
 import { CsrfErrors } from '@/shared/model/Errors';
