+++ src/server/routes/verifyEmail.ts
@@ -10,8 +10,8 @@ import { renderer } from '@/server/lib/renderer';
 import { consentPages } from '@/server/routes/consents';
 import { read as getUser } from '@/server/lib/idapi/user';
 import { ConsentsErrors, VerifyEmailErrors } from '@/shared/model/Errors';
-import { getConfiguration } from '@/server/lib/configuration';
-import { getProfileUrl } from '@/server/lib/baseUri';
+import { getConfiguration } from '@/server/lib/getConfiguration';
+import { getProfileUrl } from '@/server/lib/getProfileUrl';
 import { getProviderForEmail } from '@/shared/lib/emailProvider';
 import { trackMetric } from '@/server/lib/AWS';
 import { Metrics } from '@/server/models/Metrics';
